Rename code folder to src
