+++ openfl-workspace/fe_tf_adversarial_cifar/plan/plan.yaml
@@ -19,7 +19,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.fecifar_inmemory.FastEstimatorCifarInMemory
+  template : src.fecifar_inmemory.FastEstimatorCifarInMemory
   settings :
     collaborator_count : 2
     data_group_name    : cifar
@@ -27,7 +27,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.fe_fgsm.FastEstimatorFGSM
+  template : src.fe_fgsm.FastEstimatorFGSM
 
 network :
   defaults : plan/defaults/network.yaml
