+++ openfl-workspace/fe_tf_adversarial_cifar/src/fecifar_inmemory.py
