+++ openfl-workspace/fe_torch_adversarial_cifar/src/fecifar_inmemory.py
