+++ openfl-workspace/keras_cnn_mnist/plan/plan.yaml
@@ -19,7 +19,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.tfmnist_inmemory.TensorFlowMNISTInMemory
+  template : src.tfmnist_inmemory.TensorFlowMNISTInMemory
   settings :
     collaborator_count : 2
     data_group_name    : mnist
@@ -27,7 +27,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.keras_cnn.KerasCNN
+  template : src.keras_cnn.KerasCNN
 
 network :
   defaults : plan/defaults/network.yaml
