+++ openfl-workspace/keras_cnn_mnist/src/__init__.py
