+++ openfl-workspace/keras_cnn_mnist/src/keras_cnn.py
