+++ openfl-workspace/keras_cnn_mnist/src/mnist_utils.py
