+++ openfl-workspace/keras_cnn_mnist/src/tfmnist_inmemory.py
