+++ openfl-workspace/keras_cnn_with_compression/plan/plan.yaml
@@ -21,7 +21,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.tfmnist_inmemory.TensorFlowMNISTInMemory
+  template : src.tfmnist_inmemory.TensorFlowMNISTInMemory
   settings :
     collaborator_count : 2
     data_group_name    : mnist
@@ -29,7 +29,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.keras_cnn.KerasCNN
+  template : src.keras_cnn.KerasCNN
 
 network :
   defaults : plan/defaults/network.yaml
