+++ openfl-workspace/keras_cnn_with_compression/src/__init__.py
