+++ openfl-workspace/keras_cnn_with_compression/src/keras_cnn.py
