+++ openfl-workspace/keras_cnn_with_compression/src/mnist_utils.py
