+++ openfl-workspace/keras_cnn_with_compression/src/tfmnist_inmemory.py
