+++ openfl-workspace/keras_nlp/plan/plan.yaml
@@ -21,7 +21,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.nlp_dataloader.NLPDataLoader
+  template : src.nlp_dataloader.NLPDataLoader
   settings :
     collaborator_count : 2
     batch_size         : 64
@@ -30,7 +30,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.nlp_taskrunner.KerasNLP
+  template : src.nlp_taskrunner.KerasNLP
   settings :
     latent_dim : 256
