+++ openfl-workspace/tf_2dunet/plan/plan.yaml
@@ -19,7 +19,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.tfbrats_inmemory.TensorFlowBratsInMemory
+  template : src.tfbrats_inmemory.TensorFlowBratsInMemory
   settings :
     batch_size: 64
     percent_train: 0.8
@@ -28,7 +28,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.tf_2dunet.TensorFlow2DUNet
+  template : src.tf_2dunet.TensorFlow2DUNet
 
 network :
   defaults : plan/defaults/network.yaml
