+++ openfl-workspace/tf_2dunet/src/tfbrats_inmemory.py
