+++ openfl-workspace/tf_cnn_histology/src/tf_cnn.py
