+++ openfl-workspace/tf_cnn_histology/src/tfhistology_inmemory.py
