+++ openfl-workspace/torch_cnn_histology/plan/plan.yaml
@@ -18,7 +18,7 @@ collaborator :
     opt_treatment    : RESET
 
 data_loader :
-  template : code.pthistology_inmemory.PyTorchHistologyInMemory
+  template : src.pthistology_inmemory.PyTorchHistologyInMemory
   settings :
     collaborator_count : 2
     data_group_name    : histology
@@ -26,7 +26,7 @@ data_loader :
 
 task_runner:
   defaults : plan/defaults/task_runner.yaml
-  template: code.pt_cnn.PyTorchCNN
+  template: src.pt_cnn.PyTorchCNN
 
 network:
   defaults: plan/defaults/network.yaml
