+++ openfl-workspace/torch_cnn_histology/src/histology_utils.py
