+++ openfl-workspace/torch_cnn_histology/src/pt_cnn.py
