+++ openfl-workspace/torch_cnn_histology/src/pthistology_inmemory.py
