+++ openfl-workspace/torch_cnn_histology/src/requirements.txt
