+++ openfl-workspace/torch_cnn_mnist/plan/plan.yaml
@@ -19,7 +19,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.ptmnist_inmemory.PyTorchMNISTInMemory
+  template : src.ptmnist_inmemory.PyTorchMNISTInMemory
   settings :
     collaborator_count : 2
     data_group_name    : mnist
@@ -27,7 +27,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.pt_cnn.PyTorchCNN
+  template : src.pt_cnn.PyTorchCNN
 
 network :
   defaults : plan/defaults/network.yaml
