+++ openfl-workspace/torch_cnn_mnist/src/mnist_utils.py
