+++ openfl-workspace/torch_cnn_mnist/src/pt_cnn.py
