+++ openfl-workspace/torch_cnn_mnist/src/ptmnist_inmemory.py
