+++ openfl-workspace/torch_unet_kvasir/plan/plan.yaml
@@ -21,7 +21,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : code.data_loader.PyTorchKvasirDataLoader
+  template : src.data_loader.PyTorchKvasirDataLoader
   settings :
     collaborator_count : 2
     data_group_name    : kvasir
@@ -29,7 +29,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : code.fed_unet_runner.PyTorchFederatedUnet
+  template : src.fed_unet_runner.PyTorchFederatedUnet
   settings :
     n_channels : 3
     n_classes  : 1
